+++ src/app/app.module.ts
@@ -33,6 +33,12 @@ import {environment} from '../environments/environment';
 import {PdfCreatorService} from './services/pdf-creator.service';
 import {FormsModule, ReactiveFormsModule} from '@angular/forms';
 import { ToastContainerComponent } from './widgets/toast-container/toast-container.component';
+import {MatRadioModule} from '@angular/material/radio';
+import {MatProgressSpinnerModule} from '@angular/material/progress-spinner';
+import {MatCheckboxModule} from '@angular/material/checkbox';
+import {MatProgressBarModule} from '@angular/material/progress-bar';
+import { DialogsModule } from '@progress/kendo-angular-dialog';
+
 
 
 @NgModule({
@@ -48,50 +54,55 @@ import { ToastContainerComponent } from './widgets/toast-container/toast-contain
     SettingsComponent,
     ToastContainerComponent,
   ],
-    imports: [
-        BrowserModule,
-        AppRoutingModule,
-        BrowserAnimationsModule,
-        RouterModule.forRoot(
-            [
-                {path: '', component: LoginComponent},
-                {path: 'settings', component: SettingsComponent},
-                {path: 'signup', component: SignUpComponent},
-                {
-                    path: 'samples/overview',
-                    component: OverviewComponent,
-                    canActivate: [AngularFireAuthGuard]
-                },
-                {
-                    path: 'samples/import',
-                    component: ImportCsvComponent,
-                    canActivate: [AngularFireAuthGuard]
-                },
-                {
-                    path: 'samples/new',
-                    component: EditSampleComponent,
-                    canActivate: [AngularFireAuthGuard]
-                },
-                {
-                    path: 'samples/:id',
-                    component: EditSampleComponent,
-                    canActivate: [AngularFireAuthGuard]
-                },
-            ],
-        ),
-        NgbModule,
-        MatFormFieldModule,
-        MatInputModule,
-        MatButtonModule,
-        AngularFireModule.initializeApp(secrets),
-        InputsModule,
-        LabelModule,
-        GridModule,
-        MatCardModule,
-        MatIconModule,
-        ReactiveFormsModule,
-        FormsModule,
-    ],
+  imports: [
+    BrowserModule,
+    AppRoutingModule,
+    BrowserAnimationsModule,
+    RouterModule.forRoot(
+      [
+        {path: '', component: LoginComponent},
+        {path: 'settings', component: SettingsComponent},
+        {path: 'signup', component: SignUpComponent},
+        {
+          path: 'samples/overview',
+          component: OverviewComponent,
+          // canActivate: [AngularFireAuthGuard]
+        },
+        {
+          path: 'samples/import',
+          component: ImportCsvComponent,
+          // canActivate: [AngularFireAuthGuard]
+        },
+        {
+          path: 'samples/new',
+          component: EditSampleComponent,
+          // canActivate: [AngularFireAuthGuard]
+        },
+        {
+          path: 'samples/:id',
+          component: EditSampleComponent,
+          // canActivate: [AngularFireAuthGuard]
+        },
+      ],
+    ),
+    NgbModule,
+    MatFormFieldModule,
+    MatInputModule,
+    MatButtonModule,
+    AngularFireModule.initializeApp(secrets),
+    InputsModule,
+    LabelModule,
+    GridModule,
+    MatCardModule,
+    MatIconModule,
+    ReactiveFormsModule,
+    FormsModule,
+    MatRadioModule,
+    MatProgressSpinnerModule,
+    MatCheckboxModule,
+    MatProgressBarModule,
+    DialogsModule,
+  ],
   providers: [
     AngularFireAuth,
     AngularFireDatabase,